front-end-setup

+++ backend/controllers/goalsController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalsModel");
+const User = require("../models/userModel");
 
 // @des Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json({ goals });
 });
 
@@ -19,6 +20,7 @@ const setGoal = async (req, res) => {
   }
 
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
   res.status(200).json(goal);
@@ -32,6 +34,19 @@ const updateGoal = async (req, res) => {
 
   if (!goal) {
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+  // check of user
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  // Make user the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user Not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -48,6 +63,19 @@ const deleteGoal = async (req, res) => {
 
   if (!goal) {
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+  // check of user
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  // Make user the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401).json({ message: "user not authorized" });
+    throw new Error("user Not authorized");
   }
 
   const removeGoal = await Goal.findByIdAndRemove(req.params.id);
